Move a unit test
